% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusioMR-package.R
\docType{package}
\name{FusioMR-package}
\alias{FusioMR-package}
\alias{FusioMR}
\title{FusioMR: Flexible, Unified and Versatile Mendelian Randomization}
\description{
A Bayesian Mendelian Randomization framework for causal effect estimation from GWAS summary statistics. Supports single and multiple exposure/outcome analyses with optional correlated horizontal pleiotropy.

A Bayesian Mendelian Randomization framework for causal effect
estimation from GWAS summary statistics. Main entry:
\code{\link{fusiomr}}.
}
\author{
\strong{Maintainer}: Bowei Kang \email{kbw@uchicago.edu}

Authors:
\itemize{
  \item Sihao Feng \email{sfeng56@uchicago.edu}
}

}
\keyword{internal}
