% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_empirical_ci}
\alias{get_empirical_ci}
\title{Calculate empirical credible interval}
\usage{
get_empirical_ci(estimates, alpha = 0.05)
}
\arguments{
\item{estimates}{Vector of post estimates}

\item{alpha}{Significance level (default: 0.05)}
}
\value{
Vector of length 2 with lower and upper bounds
}
\description{
Calculate empirical credible interval
}
