% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_memo.R
\name{get_post_burnin_res}
\alias{get_post_burnin_res}
\title{Drop burnin proportion}
\usage{
get_post_burnin_res(res, niter, burnin_prop)
}
\arguments{
\item{res}{A list containing MCMC results with components:
\itemize{
  \item \code{pst_tk}: Matrix of posterior samples for joint probabilities.
  \item \code{beta_1_tk}: Vector of posterior samples for causal effect of outcome1
  \item \code{beta_2_tk}: Vector of posterior samples for causal effect of outcome2
  \item \code{eta_1_tk}: Vector of posterior samples for eta1
  \item \code{eta_2_tk}: Vector of posterior samples for eta2
  \item \code{alpha_1_tk}: Vector of posterior samples for alpha1
  \item \code{alpha_2_tk}: Vector of posterior samples for alpha2
}}

\item{niter}{An integer represents the total number of MCMC iterations.}

\item{burnin_prop}{A number between 0 and 1 representing burn-in proportion.}
}
\value{
A list with the same structure as \code{res}, containing only post-burnin samples.
}
\description{
Removes burn-in samples from MCMC results to obtain post-burnin posterior samples
for subsequent analysis.
}
