% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_sel_idx}
\alias{get_sel_idx}
\title{Get indices for selected IVs}
\usage{
get_sel_idx(b_exp, se_exp, p_threshold)
}
\arguments{
\item{b_exp}{Effect estimates for IV-exposure associations}

\item{se_exp}{Standard errors for IV-exposure associations}

\item{p_threshold}{P-value threshold for selection}
}
\value{
Logical vector indicating selected IVs
}
\description{
Get indices for selected IVs
}
