% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_summary}
\alias{get_summary}
\title{Get summary statistics from posterior estimates}
\usage{
get_summary(estimates)
}
\arguments{
\item{estimates}{Vector of posterior MCMC estimations}
}
\value{
List with beta_est, beta_se, beta_pval, ci_emp, and ci_normal
}
\description{
Get summary statistics from posterior estimates
}
