% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_init.R
\name{init_setup_seso_with_chp}
\alias{init_setup_seso_with_chp}
\title{Initialize MCMC trace for seso_with_chp}
\usage{
init_setup_seso_with_chp(
  niter,
  K,
  alpha_init,
  beta_init,
  sigma_gamma_init,
  sigma_theta_init,
  q_init = 0.1
)
}
\arguments{
\item{niter}{Number of MCMC iterations.}

\item{K}{Number of selected IVs.}

\item{alpha_init}{Starting value for the CHP magnitude alpha.}

\item{beta_init}{Starting value for the causal effect beta.}

\item{sigma_gamma_init}{Starting SD for gamma.}

\item{sigma_theta_init}{Starting SD for theta.}

\item{q_init}{Starting value for the CHP proportion q in [0, 1].}
}
\value{
A list of pre-allocated traces to pass to the C++ sampler.
}
\description{
Allocates trace storage for the seso_with_chp model, which adds
alpha (CHP magnitude), eta (CHP indicators), and q (CHP proportion)
on top of the seso_uhp_only parameters.
}
\keyword{internal}
