% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print_summary}
\alias{print_summary}
\title{Print summary results}
\usage{
print_summary(res_summary)
}
\arguments{
\item{res_summary}{Result summary object from get_summary()}
}
\description{
Print summary results
}
