% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusioMR_package.R
\docType{package}
\name{FusioMR-package}
\alias{FusioMR-package}
\alias{FusioMR}
\title{FusioMR: Fusion Mendelian Randomization Framework}
\description{
Implements a flexible, unified, and versatile framework for Mendelian Randomization analysis using Bayesian approaches with Gibbs sampling. Provides functions for instrumental variable selection with provided p_value and causal effect estimation from GWAS summary statistics.

The FusioMR package implements FusioMR framework for Mendelian Randomization
analysis using Bayesian approaches with Gibbs sampling. It provides
functions for instrumental variable selection and causal effect estimation
from GWAS summary statistics.
}
\section{Main functions}{

\describe{
  \item{\code{\link{fusiomr}}}{Main function for causal effect estimation with given summary statistics}
  \item{\code{\link{summary_stats_selected}}}{Implement instrumental variable selection with given p-value}
}
}

\section{Types of framework}{

\describe{
  \item{seso_uhp_only}{Single exposure, single outcome with uncorrelated horizontal pleiotropy only}
  \item{seso_nohp}{Single exposure, single outcome with no horizontal pleiotropy}
  \item{multi_uhp_only}{Multiple exposures with uncorrelated horizontal pleiotropy (not yet implemented)}
  \item{multi_nohp}{Multiple exposures with no horizontal pleiotropy (not yet implemented)}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kangbw702/FusioMR}
  \item Report bugs at \url{https://github.com/kangbw702/FusioMR/issues}
}

}
\author{
\strong{Maintainer}: Sihao Feng \email{sfeng56@uchicago.edu}

}
\keyword{internal}
